use CONTINUE instead of ACCEPT
When running multiple milters it is better to use CONTINUE as that will allow other milters to continue to process. ACCEPT or REJECT is a shortcut that will prevent other milters to do things, even if we mark this message as spam, for delivery in the spam folder, other milters may still REJECT it for other reasons, which means not deliver.